requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
dj-rest-auth==2.1.4
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.9
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
